requirements.txt
flask==3.0.3
flask-mysqldb==1.0.1
qrcode==7.4.2
Pillow==10.4.0
stripe==9.12.0
python-dotenv==1.0.1
werkzeug==3.0.4
